Save scroll on sheet updates
