+++ scripts/hooks/ready.js
@@ -1,3 +1,5 @@
+import {migrateWorld} from "../common/migration.js"
+
 export default function() {
     Hooks.once("ready", () => {
         migrateWorld();
